@@ -1,4 +1,5 @@
 using SharedEntities;
+using ZavaFoundryAgentsProvider;
 
 namespace AgentsCatalogService;
 
@@ -8,55 +9,18 @@ public static class AgentCatalog
 
     public static void InitAgentList(IConfiguration config)
     {
-        Agents.Add(new AvailableAgent
+        Agents.Clear();
+        foreach (AgentNamesProvider.AgentName agentEnum in System.Enum.GetValues(typeof(AgentNamesProvider.AgentName)))
         {
-            AgentId = "customerinformationagentid",
-            AgentCnnStringId = "customerinformationagentid",
-            AgentName = "Customer Information Agent",
-            Description = "Manages customer information and provides personalized recommendations"
-        });
-        Agents.Add(new AvailableAgent
-        {
-            AgentId = "inventoryagentid",
-            AgentCnnStringId = "inventoryagentid",
-            AgentName = "Inventory Agent",
-            Description = "Searches product inventory and provides availability information"
-        });
-        Agents.Add(new AvailableAgent
-        {
-            AgentId = "locationserviceagentid",
-            AgentCnnStringId = "locationserviceagentid",
-            AgentName = "Location Service Agent",
-            Description = "Provides location-based services and store information"
-        });
-        Agents.Add(new AvailableAgent
-        {
-            AgentId = "navigationagentid",
-            AgentCnnStringId = "navigationagentid",
-            AgentName = "Navigation Agent",
-            Description = "Provides navigation and routing assistance within stores"
-        });
-        Agents.Add(new AvailableAgent
-        {
-            AgentId = "photoanalyzeragentid",
-            AgentCnnStringId = "photoanalyzeragentid",
-            AgentName = "Photo Analysis Agent",
-            Description = "Analyzes photos and identifies materials and project requirements"
-        });
-        Agents.Add(new AvailableAgent
-        {
-            AgentId = "productmatchmakingagentid",
-            AgentCnnStringId = "productmatchmakingagentid",
-            AgentName = "Product Matchmaking Agent",
-            Description = "Matches products to customer needs and project requirements"
-        });
-        Agents.Add(new AvailableAgent
-        {
-            AgentId = "toolreasoningagentid",
-            AgentCnnStringId = "toolreasoningagentid",
-            AgentName = "Tool Reasoning Agent",
-            Description = "Provides reasoning for tool recommendations based on DIY projects"
-        });
+            var agentName = AgentNamesProvider.GetAgentName(agentEnum);
+            Agents.Add(new AvailableAgent
+            {
+                AgentId = agentName.ToLowerInvariant() + "id",
+                AgentCnnStringId = agentName.ToLowerInvariant() + "id",
+                AgentName = agentName.Replace("Agent", " Agent"),
+                Description = $"Agent for {agentName.Replace("Agent", " agent").Replace("Matchmaking", "matchmaking").Replace("Reasoning", "reasoning").Replace("Information", "information").Replace("PhotoAnalyzer", "photo analysis").Replace("ProductSearch", "product search").Replace("LocationService", "location service").Replace("Navigation", "navigation").Replace("Inventory", "inventory").ToLower()}"
+            });
+        }
         foreach (var agent in Agents)
         {
             var agentId = config.GetConnectionString(agent.AgentCnnStringId);
