@@ -1,6 +1,6 @@
 using AgentsCatalogService;
-using ZavaAIFoundrySKAgentsProvider;
-using ZavaAgentFxAgentsProvider;
+using ZavaMAFAgentsProvider;
+using ZavaFoundryAgentsProvider;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -12,20 +12,14 @@ builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
 
-// Add AI Foundry Agent Provider for agent testing
-builder.Services.AddSingleton<AIFoundryAgentProvider>(sp =>
+/********************************************************/
+// The following code registers the agent providers for the Microsoft Foundry project.  
+var microsoftFoundryProjectConnection = builder.Configuration.GetConnectionString("microsoftfoundryproject");
+builder.Services.AddSingleton(sp =>
 {
-    var config = sp.GetService<IConfiguration>();
-    var aiFoundryProjectConnection = config?.GetConnectionString("aifoundryproject") ?? "";
-    return new AIFoundryAgentProvider(aiFoundryProjectConnection, "");
+    return new MAFAgentProvider(microsoftFoundryProjectConnection!);
 });
-
-builder.Services.AddSingleton<AgentFxAgentProvider>(sp =>
-{
-    var config = sp.GetService<IConfiguration>();
-    var aiFoundryProjectConnection = config?.GetConnectionString("aifoundryproject") ?? "";
-    return new AgentFxAgentProvider(aiFoundryProjectConnection);
-});
+/********************************************************/
 
 var app = builder.Build();
 
