@@ -1,73 +1,32 @@
-#pragma warning disable SKEXP0110
 using Microsoft.Agents.AI;
-using Microsoft.Extensions.AI;
-using Microsoft.SemanticKernel.Agents.AzureAI;
-using Microsoft.SemanticKernel.ChatCompletion;
-using ZavaAIFoundrySKAgentsProvider;
-using ZavaAgentFxAgentsProvider;
-using ZavaSemanticKernelProvider;
+using ZavaFoundryAgentsProvider;
+using ZavaMAFAgentsProvider;
 
 var builder = WebApplication.CreateBuilder(args);
 
 builder.AddServiceDefaults();
 
-// Add services to the container.
+// Add services to the container
 builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
-builder.Services.AddSingleton(sp =>
-{
-    var config = sp.GetService<IConfiguration>();
-    var aiFoundryConnection = config.GetConnectionString("aifoundry");
-    var chatDeploymentName = config["AI_ChatDeploymentName"] ?? "gpt-5-mini";
-    return new SemanticKernelProvider(aiFoundryConnection, chatDeploymentName);
-});
+// Register MAFAgentProvider for Microsoft Foundry integration
+var microsoftFoundryProjectConnection = builder.Configuration.GetConnectionString("microsoftfoundryproject");
+builder.Services.AddSingleton(_ => new MAFAgentProvider(microsoftFoundryProjectConnection!));
 
-builder.Services.AddSingleton<IChatClient>(sp =>
-{
-    var skProvider = sp.GetService<SemanticKernelProvider>();
-    var kernel = skProvider.GetKernel();
-    return kernel.GetRequiredService<IChatCompletionService>().AsChatClient();
-});
-
-// Register both agent providers for dual framework support
-builder.Services.AddSingleton(sp =>
-{
-    var config = sp.GetService<IConfiguration>();
-    var aiFoundryProjectConnection = config.GetConnectionString("aifoundryproject");
-    var agentId = config.GetConnectionString("customerinformationagentid");
-    return new AIFoundryAgentProvider(aiFoundryProjectConnection, agentId);
-});
-
-builder.Services.AddSingleton(sp =>
-{
-    var config = sp.GetService<IConfiguration>();
-    var aiFoundryProjectConnection = config.GetConnectionString("aifoundryproject");
-    return new AgentFxAgentProvider(aiFoundryProjectConnection!);
-});
-
-builder.Services.AddSingleton<AzureAIAgent>(sp =>
-{
-    var config = sp.GetService<IConfiguration>();
-    var agentId = config.GetConnectionString("customerinformationagentid");
-    var aiFoundryAgentProvider = sp.GetService<AIFoundryAgentProvider>();
-    return aiFoundryAgentProvider.CreateAzureAIAgent(agentId);
-});
-
+// Register the CustomerInformationAgent from Microsoft Foundry
 builder.Services.AddSingleton<AIAgent>(sp =>
 {
-    var config = sp.GetService<IConfiguration>();
-    var agentId = config.GetConnectionString("customerinformationagentid");
-    var agentFxProvider = sp.GetService<AgentFxAgentProvider>();
-    return agentFxProvider.GetAIAgent(agentId);
+    var agentId = AgentNamesProvider.GetAgentName(AgentNamesProvider.AgentName.CustomerInformationAgent);
+    var provider = sp.GetRequiredService<MAFAgentProvider>();
+    return provider.GetAIAgent(agentId);
 });
 
 var app = builder.Build();
 
 app.MapDefaultEndpoints();
 
-// Configure the HTTP request pipeline.
 if (app.Environment.IsDevelopment())
 {
     app.UseSwagger();
