@@ -4,7 +4,7 @@ using SharedEntities;
 namespace MultiAgentDemo.Services;
 
 /// <summary>
-/// Group Chat orchestration - all agents participate in a group conversation, coordinated by a group manager.
+/// Group Chat orchestration - all agents participate in a group conversation, coordinated by a manager.
 /// Use Case: Brainstorming, collaborative problem solving, consensus building.
 /// </summary>
 public class GroupChatOrchestrationService : IAgentOrchestrationService
@@ -29,6 +29,7 @@ public class GroupChatOrchestrationService : IAgentOrchestrationService
         _navigationAgentService = navigationAgentService;
     }
 
+    /// <inheritdoc />
     public async Task<MultiAgentResponse> ExecuteAsync(MultiAgentRequest request)
     {
         var orchestrationId = Guid.NewGuid().ToString();
@@ -37,52 +38,28 @@ public class GroupChatOrchestrationService : IAgentOrchestrationService
         var steps = new List<AgentStep>();
         var conversationContext = new List<string>();
 
-        // Group Manager initiates the conversation
-        var managerStep = CreateManagerStep("Group Manager", "Initiate discussion", 
-            $"Welcome to the group discussion about '{request.ProductQuery}'. Let's collaborate to help the customer.", DateTime.UtcNow);
+        // Initiate the group discussion
+        var managerStep = CreateManagerStep("Initiate discussion",
+            $"Welcome to the group discussion about '{request.ProductQuery}'. Let's collaborate to help the customer.");
         steps.Add(managerStep);
         conversationContext.Add($"Manager: {managerStep.Result}");
 
         // Round 1: Initial thoughts from all agents
-        var inventoryStep = await RunInventoryAgentInGroupAsync(request.ProductQuery, conversationContext, 1);
-        steps.Add(inventoryStep);
-        conversationContext.Add($"Inventory: {inventoryStep.Result}");
-
-        var matchmakingStep = await RunMatchmakingAgentInGroupAsync(request.ProductQuery, request.UserId, conversationContext, 1);
-        steps.Add(matchmakingStep);
-        conversationContext.Add($"Matchmaking: {matchmakingStep.Result}");
-
-        var locationStep = await RunLocationAgentInGroupAsync(request.ProductQuery, conversationContext, 1);
-        steps.Add(locationStep);
-        conversationContext.Add($"Location: {locationStep.Result}");
+        await ExecuteRoundOneAsync(request, steps, conversationContext);
 
         // Manager summarizes round 1
-        var managerSummary1 = CreateManagerStep("Group Manager", "Summarize Round 1", 
-            "Great initial insights! Inventory found products, Matchmaking identified alternatives, Location provided coordinates. Let's build consensus.", DateTime.UtcNow);
-        steps.Add(managerSummary1);
-        conversationContext.Add($"Manager: {managerSummary1.Result}");
+        var summary = CreateManagerStep("Summarize Round 1",
+            "Great initial insights! Inventory found products, Matchmaking identified alternatives, Location provided coordinates. Let's build consensus.");
+        steps.Add(summary);
+        conversationContext.Add($"Manager: {summary.Result}");
 
         // Round 2: Agents respond to each other's insights
-        var inventoryResponse = await RunInventoryAgentInGroupAsync(request.ProductQuery, conversationContext, 2);
-        steps.Add(inventoryResponse);
-        conversationContext.Add($"Inventory: {inventoryResponse.Result}");
+        await ExecuteRoundTwoAsync(request, steps, conversationContext);
 
-        var matchmakingResponse = await RunMatchmakingAgentInGroupAsync(request.ProductQuery, request.UserId, conversationContext, 2);
-        steps.Add(matchmakingResponse);
-        conversationContext.Add($"Matchmaking: {matchmakingResponse.Result}");
-
-        // Navigation agent joins if location is provided
-        if (request.Location != null)
-        {
-            var navigationStep = await RunNavigationAgentInGroupAsync(request.Location, request.ProductQuery, conversationContext);
-            steps.Add(navigationStep);
-            conversationContext.Add($"Navigation: {navigationStep.Result}");
-        }
-
-        // Manager concludes the discussion
-        var managerConclusion = CreateManagerStep("Group Manager", "Conclude discussion", 
-            "Excellent collaboration! We've reached consensus on the best customer solution through group discussion.", DateTime.UtcNow);
-        steps.Add(managerConclusion);
+        // Manager concludes
+        var conclusion = CreateManagerStep("Conclude discussion",
+            "Excellent collaboration! We've reached consensus on the best customer solution through group discussion.");
+        steps.Add(conclusion);
 
         NavigationInstructions? navigation = null;
         if (request.Location != null)
@@ -90,154 +67,162 @@ public class GroupChatOrchestrationService : IAgentOrchestrationService
             navigation = await GenerateNavigationInstructionsAsync(request.Location, request.ProductQuery);
         }
 
-        // Generate mock alternatives for UI compatibility
-        var alternatives = StepsProcessor.GenerateDefaultProductAlternatives();
-
         return new MultiAgentResponse
         {
             OrchestrationId = orchestrationId,
             OrchestationType = OrchestrationType.GroupChat,
-            OrchestrationDescription = "Agents participated in a collaborative group chat coordinated by a group manager, with multiple rounds of discussion to build consensus and share insights.",
+            OrchestrationDescription = "Agents participated in a collaborative group chat with multiple rounds of discussion to build consensus.",
             Steps = steps.ToArray(),
-            Alternatives = alternatives,
+            Alternatives = StepsProcessor.GenerateDefaultProductAlternatives(),
             NavigationInstructions = navigation
         };
     }
 
-    private AgentStep CreateManagerStep(string agent, string action, string result, DateTime timestamp)
+    private async Task ExecuteRoundOneAsync(MultiAgentRequest request, List<AgentStep> steps, List<string> context)
     {
-        return new AgentStep
-        {
-            Agent = agent,
-            Action = action,
-            Result = result,
-            Timestamp = timestamp
-        };
+        var inventoryStep = await ExecuteInventoryAgentInGroupAsync(request.ProductQuery, 1);
+        steps.Add(inventoryStep);
+        context.Add($"Inventory: {inventoryStep.Result}");
+
+        var matchmakingStep = await ExecuteMatchmakingAgentInGroupAsync(request.ProductQuery, request.UserId, 1);
+        steps.Add(matchmakingStep);
+        context.Add($"Matchmaking: {matchmakingStep.Result}");
+
+        var locationStep = await ExecuteLocationAgentInGroupAsync(request.ProductQuery, 1);
+        steps.Add(locationStep);
+        context.Add($"Location: {locationStep.Result}");
     }
 
-    private async Task<AgentStep> RunInventoryAgentInGroupAsync(string productQuery, List<string> context, int round)
+    private async Task ExecuteRoundTwoAsync(MultiAgentRequest request, List<AgentStep> steps, List<string> context)
+    {
+        var inventoryResponse = await ExecuteInventoryAgentInGroupAsync(request.ProductQuery, 2);
+        steps.Add(inventoryResponse);
+        context.Add($"Inventory: {inventoryResponse.Result}");
+
+        var matchmakingResponse = await ExecuteMatchmakingAgentInGroupAsync(request.ProductQuery, request.UserId, 2);
+        steps.Add(matchmakingResponse);
+        context.Add($"Matchmaking: {matchmakingResponse.Result}");
+
+        if (request.Location != null)
+        {
+            var navigationStep = await ExecuteNavigationAgentInGroupAsync(request.Location, request.ProductQuery);
+            steps.Add(navigationStep);
+            context.Add($"Navigation: {navigationStep.Result}");
+        }
+    }
+
+    private static AgentStep CreateManagerStep(string action, string result) => new()
+    {
+        Agent = "Group Manager",
+        Action = action,
+        Result = result,
+        Timestamp = DateTime.UtcNow
+    };
+
+    private async Task<AgentStep> ExecuteInventoryAgentInGroupAsync(string productQuery, int round)
     {
         try
         {
             var result = await _inventoryAgentService.SearchProductsAsync(productQuery);
-            var names = result?.ProductsFound?.Select(p => p.Name) ?? Enumerable.Empty<string>();
-            
-            string response;
-            if (round == 1)
-            {
-                response = $"Group discussion: I found {result?.TotalCount ?? 0} products for '{productQuery}': {string.Join(", ", names)}. What do others think?";
-            }
-            else
-            {
-                response = $"Following up on the discussion: I can confirm stock levels and suggest cross-checking with the alternatives mentioned by Matchmaking.";
-            }
+            var productNames = result?.ProductsFound?.Select(p => p.Name) ?? [];
 
-            return new AgentStep { Agent = "InventoryAgent", Action = $"Group discussion Round {round}", Result = response, Timestamp = DateTime.UtcNow };
+            var response = round == 1
+                ? $"Group discussion: I found {result?.TotalCount ?? 0} products for '{productQuery}': {string.Join(", ", productNames)}. What do others think?"
+                : "Following up: I can confirm stock levels and suggest cross-checking with the alternatives mentioned by Matchmaking.";
+
+            return CreateStep("InventoryAgent", $"Group discussion Round {round}", response);
         }
         catch (Exception ex)
         {
             _logger.LogWarning(ex, "Inventory agent failed in group chat");
-            return new AgentStep { Agent = "InventoryAgent", Action = $"Group discussion Round {round}", Result = "Having technical issues, will follow up", Timestamp = DateTime.UtcNow };
+            return CreateStep("InventoryAgent", $"Group discussion Round {round}", "Having technical issues, will follow up");
         }
     }
 
-    private async Task<AgentStep> RunMatchmakingAgentInGroupAsync(string productQuery, string userId, List<string> context, int round)
+    private async Task<AgentStep> ExecuteMatchmakingAgentInGroupAsync(string productQuery, string userId, int round)
     {
         try
         {
             var result = await _matchmakingAgentService.FindAlternativesAsync(productQuery, userId);
             var count = result?.Alternatives?.Length ?? 0;
-            
-            string response;
-            if (round == 1)
-            {
-                response = $"Group input: I've identified {count} alternatives for '{productQuery}'. These could complement what Inventory found.";
-            }
-            else
-            {
-                response = $"Building on Location's findings: I can match alternatives to specific aisles they mentioned. Great teamwork!";
-            }
 
-            return new AgentStep { Agent = "MatchmakingAgent", Action = $"Group discussion Round {round}", Result = response, Timestamp = DateTime.UtcNow };
+            var response = round == 1
+                ? $"Group input: I've identified {count} alternatives for '{productQuery}'. These could complement what Inventory found."
+                : "Building on Location's findings: I can match alternatives to specific aisles they mentioned. Great teamwork!";
+
+            return CreateStep("MatchmakingAgent", $"Group discussion Round {round}", response);
         }
         catch (Exception ex)
         {
             _logger.LogWarning(ex, "Matchmaking agent failed in group chat");
-            return new AgentStep { Agent = "MatchmakingAgent", Action = $"Group discussion Round {round}", Result = "Experiencing delays, will contribute next round", Timestamp = DateTime.UtcNow };
+            return CreateStep("MatchmakingAgent", $"Group discussion Round {round}", "Experiencing delays, will contribute next round");
         }
     }
 
-    private async Task<AgentStep> RunLocationAgentInGroupAsync(string productQuery, List<string> context, int round)
+    private async Task<AgentStep> ExecuteLocationAgentInGroupAsync(string productQuery, int round)
     {
         try
         {
             var result = await _locationAgentService.FindProductLocationAsync(productQuery);
-            var loc = result?.StoreLocations?.FirstOrDefault();
-            
-            string response;
-            if (round == 1)
-            {
-                response = loc != null ? 
-                    $"Group collaboration: Found '{productQuery}' in {loc.Section} Aisle {loc.Aisle}. This aligns with Inventory's findings!" : 
-                    "Group discussion: No specific location found, but Matchmaking's alternatives might help.";
-            }
-            else
-            {
-                response = "Reflecting on our discussion: I can provide detailed aisle maps to support Navigation's route planning.";
-            }
+            var location = result?.StoreLocations?.FirstOrDefault();
 
-            return new AgentStep { Agent = "LocationAgent", Action = $"Group discussion Round {round}", Result = response, Timestamp = DateTime.UtcNow };
+            var response = round == 1
+                ? location != null
+                    ? $"Group collaboration: Found '{productQuery}' in {location.Section} Aisle {location.Aisle}. This aligns with Inventory's findings!"
+                    : "Group discussion: No specific location found, but Matchmaking's alternatives might help."
+                : "Reflecting on our discussion: I can provide detailed aisle maps to support Navigation's route planning.";
+
+            return CreateStep("LocationAgent", $"Group discussion Round {round}", response);
         }
         catch (Exception ex)
         {
             _logger.LogWarning(ex, "Location agent failed in group chat");
-            return new AgentStep { Agent = "LocationAgent", Action = $"Group discussion Round {round}", Result = "Technical difficulties, deferring to group", Timestamp = DateTime.UtcNow };
+            return CreateStep("LocationAgent", $"Group discussion Round {round}", "Technical difficulties, deferring to group");
         }
     }
 
-    private async Task<AgentStep> RunNavigationAgentInGroupAsync(Location? location, string productQuery, List<string> context)
+    private async Task<AgentStep> ExecuteNavigationAgentInGroupAsync(Location? location, string productQuery)
     {
+        if (location == null)
+        {
+            return CreateStep("NavigationAgent", "Join group discussion", "Happy to help but need customer start location");
+        }
+
         try
         {
-            if (location == null) return new AgentStep { Agent = "NavigationAgent", Action = "Join group discussion", Result = "Happy to help but need customer start location", Timestamp = DateTime.UtcNow };
-            
-            var dest = new Location { Lat = 0, Lon = 0 };
-            var nav = await _navigationAgentService.GenerateDirectionsAsync(location, dest);
-            var steps = nav?.Steps?.Length ?? 0;
-            var response = $"Joining the discussion: Based on Location's coordinates and Inventory's findings, I can provide {steps} navigation steps!";
-            
-            return new AgentStep { Agent = "NavigationAgent", Action = "Join group discussion", Result = response, Timestamp = DateTime.UtcNow };
+            var destination = new Location { Lat = 0, Lon = 0 };
+            var nav = await _navigationAgentService.GenerateDirectionsAsync(location, destination);
+            var stepCount = nav?.Steps?.Length ?? 0;
+            var response = $"Joining the discussion: Based on Location's coordinates and Inventory's findings, I can provide {stepCount} navigation steps!";
+
+            return CreateStep("NavigationAgent", "Join group discussion", response);
         }
         catch (Exception ex)
         {
             _logger.LogWarning(ex, "Navigation agent failed in group chat");
-            return new AgentStep { Agent = "NavigationAgent", Action = "Join group discussion", Result = "Working on route calculation, great group effort so far!", Timestamp = DateTime.UtcNow };
+            return CreateStep("NavigationAgent", "Join group discussion", "Working on route calculation, great group effort so far!");
         }
     }
 
-    private async Task<NavigationInstructions> GenerateNavigationInstructionsAsync(Location? location, string productQuery)
+    private async Task<NavigationInstructions> GenerateNavigationInstructionsAsync(Location location, string productQuery)
     {
-        if (location == null) return new NavigationInstructions { Steps = Array.Empty<NavigationStep>(), StartLocation = string.Empty, EstimatedTime = string.Empty };
-        var dest = new Location { Lat = 0, Lon = 0 };
         try
         {
-            return await _navigationAgentService.GenerateDirectionsAsync(location, dest);
+            var destination = new Location { Lat = 0, Lon = 0 };
+            return await _navigationAgentService.GenerateDirectionsAsync(location, destination);
         }
         catch (Exception ex)
         {
             _logger.LogWarning(ex, "GenerateNavigationInstructions failed");
-            return new NavigationInstructions { Steps = new[] { new NavigationStep { Direction = "General", Description = $"Head to the area where {productQuery} is typically located", Landmark = new NavigationLandmark { Description = "General area" } } }, StartLocation = string.Empty, EstimatedTime = string.Empty };
+            return StepsProcessor.CreateDefaultNavigationInstructions(location, productQuery);
         }
     }
 
-    private async Task<ProductAlternative[]> GenerateProductAlternativesAsync(string productQuery)
+    private static AgentStep CreateStep(string agent, string action, string result) => new()
     {
-        await Task.Delay(10);
-        return new[]
-        {
-            new ProductAlternative { Name = $"Group-recommended {productQuery}", Sku = "GRP-" + productQuery.Replace(" ", "").ToUpper(), Price = 119.99m, InStock = true, Location = "Aisle 4", Aisle = 4, Section = "A" },
-            new ProductAlternative { Name = $"Consensus {productQuery}", Sku = "CON-" + productQuery.Replace(" ", "").ToUpper(), Price = 69.99m, InStock = true, Location = "Aisle 6", Aisle = 6, Section = "B" },
-            new ProductAlternative { Name = $"Team-selected {productQuery}", Sku = "TEAM-" + productQuery.Replace(" ", "").ToUpper(), Price = 39.99m, InStock = false, Location = "Aisle 10", Aisle = 10, Section = "C" }
-        };
-    }
+        Agent = agent,
+        Action = action,
+        Result = result,
+        Timestamp = DateTime.UtcNow
+    };
 }
