@@ -113,7 +113,7 @@ namespace Products.Tests
             }
             using (var context = new Context(_dbOptions))
             {
-                await Assert.ThrowsExceptionAsync<InvalidOperationException>(async () =>
+                await Assert.ThrowsAsync<InvalidOperationException>(async () =>
                 {
                     await ProductApiActions.DeleteProduct(40, context);
                 });
