@@ -3,6 +3,7 @@
 @using Store.Services
 @using Microsoft.AspNetCore.Components.Web
 @using SharedEntities
+@using ZavaWorkingModes
 @inject SingleAgentService SingleAgentService
 @inject ILogger<Scenario1SingleAgent> Logger
 @inject AgentFrameworkService AgentFrameworkService
@@ -19,17 +20,9 @@
             <div class="alert alert-info d-flex align-items-center" role="alert">
                 <span class="bi bi-info-circle me-2"></span>
                 <div>
-                    <strong>Current Working Mode:</strong> @currentFramework
+                    <strong>Current Working Mode:</strong> @WorkingModeProvider.GetDisplayName(currentMode)
                     <small class="ms-2 text-muted">
-                        @{
-                            string frameworkLabel = currentFramework switch
-                            {
-                                "SK" => "(Semantic Kernel)",
-                                "llm" => "(LLM Framework)",
-                                _ => "(Microsoft Agent Framework)"
-                            };
-                        }
-                        @frameworkLabel
+                        (@WorkingModeProvider.GetShortName(currentMode))
                     </small>
                 </div>
             </div>
@@ -215,11 +208,11 @@
     private string? imagePreviewUrl;
     private bool isSubmitting = false;
     private string errorMessage = string.Empty;
-    private string currentFramework = "SK";
+    private WorkingMode currentMode = WorkingModeProvider.DefaultMode;
     
     protected override async Task OnInitializedAsync()
     {
-        currentFramework = await AgentFrameworkService.GetSelectedFrameworkAsync();
+        currentMode = await AgentFrameworkService.GetSelectedModeAsync();
     }
 
     private void OnImageSelected(InputFileChangeEventArgs e)
