@@ -1,11 +1,12 @@
 using Microsoft.JSInterop;
+using ZavaWorkingModes;
 
 namespace Store.Services;
 
 public class AgentFrameworkService
 {
     private readonly IJSRuntime _jsRuntime;
-    private string? _cachedFramework;
+    private WorkingMode? _cachedMode;
 
     public AgentFrameworkService(IJSRuntime jsRuntime)
     {
@@ -14,33 +15,46 @@ public class AgentFrameworkService
 
     public async Task<string> GetSelectedFrameworkAsync()
     {
-        if (_cachedFramework != null)
+        var mode = await GetSelectedModeAsync();
+        return WorkingModeProvider.GetShortName(mode);
+    }
+
+    public async Task<WorkingMode> GetSelectedModeAsync()
+    {
+        if (_cachedMode.HasValue)
         {
-            return _cachedFramework;
+            return _cachedMode.Value;
         }
 
         try
         {
             var framework = await _jsRuntime.InvokeAsync<string>("localStorage.getItem", "agentFramework");
-            _cachedFramework = string.IsNullOrEmpty(framework) ? "agentfx" : framework.ToLowerInvariant();
-            return _cachedFramework;
+            _cachedMode = WorkingModeProvider.Parse(framework);
+            return _cachedMode.Value;
         }
         catch
         {
-            // Default to agentfx if localStorage is not available
-            _cachedFramework = "agentfx";
-            return _cachedFramework;
+            // Default to MafFoundry if localStorage is not available
+            _cachedMode = WorkingModeProvider.DefaultMode;
+            return _cachedMode.Value;
         }
     }
 
     public async Task SetSelectedFrameworkAsync(string framework)
     {
-        _cachedFramework = framework;
+        _cachedMode = WorkingModeProvider.Parse(framework);
         await _jsRuntime.InvokeVoidAsync("localStorage.setItem", "agentFramework", framework);
+    }
+
+    public async Task SetSelectedModeAsync(WorkingMode mode)
+    {
+        _cachedMode = mode;
+        var shortName = WorkingModeProvider.GetShortName(mode);
+        await _jsRuntime.InvokeVoidAsync("localStorage.setItem", "agentFramework", shortName);
     }
 
     public void ClearCache()
     {
-        _cachedFramework = null;
+        _cachedMode = null;
     }
 }
